Created two shortcuts (pyHegel_cmd, pyHegel_console) to directly execute pyHegel
They start pyHegel in a windows cmd or in a console program.
For python xy version 2.7.2.0 at least it requires disabling loading
of  pylab (-nopylab_import_all) because it is done after loading the
pyHegel file and overwrites commands like load.
That is why pylab is now loaded within pyHegel.