Fix a locking problem
The bug was using the blocking RLock acquire (by accident: forgot
to overwrite the __enter__ method)
The problem would show up when breaking into a long async (acq1.readval),
because then the lock would get acquired, but the KeyboardInterrupt
would prevent RLock from knowing about it. So on the next call,
it would try to acquire it again.

I think that is still a possibility but should be a rare event.
lock_force_release should remove the problem if it happens.