Fixed memory leak bug
When doing a loop of sweeps, with close_after, the traces would
hide but not be deleted from memory (it was waiting for a return
to the main loop, i.e. the console prompt).

I included a memorytest python file to tests the memory response
under various conditions.

There is still room for improvements: Closing a trace by pressing
the close icon does run recover the memory. The memory only becomes
freed after gc.collect().