Fix _get_async_block to not block when we already have current_thread
When we own current_thread, eventually the task is runned and that
will own the _lock_instrument. If we call it we are blocked.

With the previous (wrong) code, it would block while the task ran
and unblock once finished. So it would block after the first
async==1 and not wait at all for async==2. This would mostly work
but was wrong, and caused problems with LogicalDevice.