Fix multithreading bug with pyvisa reads (warnings)
The problem showed up when reading 2 dmm and 1 exa, with async (getasync
or async sweep). The measurement started at around 0.2 s per cycle but
eventually grew. It could more than double in an hour.

The cause was interactions between warnings.filterwarnings inserting
an element at the start of the filters list and pyvisa.visa._removefilter
recreating the filters list. Because of multithreading, (and the exa having
many reads), the insert would happen during list recreation which cause an
element to be copied twice. Therefore the list kept growing which slowed
down the list recreation (which iterates through all elemets).

Going through the list is inefficient but not really a problem as long as
the list is short (which was the case until this problem showed up).

The simplest fix was just to wrap the two functions with a lock.

If there are other frequent calls to warnings.filterwarnings they should
use the wrapped version.