logical: cleanup _cache usage, bug fix and improvements
- replaced _cache by setcache in ScalingDevice
- Main logical device get converted to _getdev.
- Sub logical device _getdev now replaced by _getdev_log
- Make getasync work properly with new _getdev and extend it
- Added basedev_as_param to FunctionWrap

The problem that needed fixing was that using get with a filename with
FunctionWrap ended up with the subdevice data in the file instead of the
converted one. This was because my get function did not handle the
filename option so it was passed on to the subdevice. The changes (converting
get to _getdev) now fixed the problem (the data saved is the
converted one now).