commands, intruments_base, logical: modified getasync to make logical devices work
Previously there was a sequencing error when using logical devices that
depended on a subdevice that was also read during async.
For example doing:
 load('foo1')
 foo1.incr._trig=True
 foo1.current._trig=True
 ll=instruments.FunctionWrap(getfunc=lambda x: x[0]/100., basedev=foo1.incr, autoget=True, basedev_as_param=True)
 getasync([foo1.current,foo1.incr, foo1.incr, ll, foo1.incr])
would not result in the expected result. The async_counter was being reset, and
the data was exctracted in a different order that it was setup.

This is now fixed. Data is always read as soon as it is available (same order everywhere).
And instruments can return immediately when they know they are not ready.