commands: correctly handle multiple sets with same dev (different options)
The problem fixed was that in sweep_multi it would used the wrong
cached value. For example for 2 dim with a device address by ch.
For v1 = (dev, dict(ch=1))
    v2 = (dev, dict(ch=2))
then sweep_multi [v1, v2],[1,100], [10, 110], 10  would produce in the data file
#dev dev
1    100
100  101
101  102
..
109  110
2    100

But now it is:
#dev dev2
1 100
1 101
...
1 110
2 100
2 101

The titles and header have also been changed to remove duplicate.
But extra_conf duplicates are not removed.